pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-harness"
version = "0.1.0"
description = "에이전트/스킬 호출 검증 및 컨텍스트 윈도우 효율성 테스트 하네스"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
agent-harness = "agent_harness.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_harness"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
